requirements.txt
streamlit
supabase
pathlib
python-dotenv
werkzeug
pandas